Correction des warnings de compilation
- Ajout des prototypes manquants dans main.c
- Correction des paramètres non utilisés avec (void)param
- Remplacement de pcap_lookupdev() déprécié par pcap_findalldevs()
- Correction de la comparaison de signes dans transport.c
- Ajout des includes manquants dans data_link.c et transport.c
- Amélioration de la gestion mémoire pour l'interface réseau

+++ transport.c
@@ -3,6 +3,7 @@
 #include <netinet/tcp.h>
 #include <netinet/udp.h>
 
+#include "transport.h"
 #include "application.h"
 
 //--------------------------------------
@@ -174,7 +175,7 @@ void tcp_view(const u_char *packet, int tcp_size) {
 	// 1 octet pour le type de l'option, 1 octets pour sa longueur et 2 pour sa valeur
 	// source est + d'infos sur https://www.iana.org/assignments/tcp-parameters/tcp-parameters.xhtml
 
-	if(tcp_size > sizeof(struct tcphdr)) {
+	if(tcp_size > (int)sizeof(struct tcphdr)) {
 	
 		printf("\t\tOptions:\n");
 		for(i=sizeof(struct tcphdr); i<tcphdr_size && packet[i] != 0x00; i++) {
